submission/submission1/requirements.txt
torch==2.10.0
torchvision==0.25.0
Pillow==12.1.1